Improved query time tracking.
Queries now record the time sent (e.g. the timestamp from google's
servers) as well as the time the query was created.

The xml parser recorded the time sent as time recieved, and the
test needed to be changed to reflect the new paradigm. Furthermore,
depending on what is important to log, the log may want to record
time from sent to response instead of time from recieve to response.