Added main class PioText.
Added an overarching class, PioText.  As a result of this, keywords
was changed to no longer function as a singleton.  Almost all other
changes are minor ones related to Keywords and Queries having slight
method signature changes.

Need to stop making these huge commits, and do them in smaller
chunks in the future.  Perhaps if I have commit timer that goes
off every 15 minutes to remind me, it would help.