Fixed tests, and made identifyKeyword() protected.
Several tests were failing because things have been changing.
They're better now.

Also, MessageHandler.identifyKeyword() is now protected, and it
is the responsibility of the implementing class to call this.

Things are pretty friendly for creation of new message handlers at
this point, as long as identify keyword is called before anything
else.  Also, data encapsulation is getting much better.