Prepped framework for deleting sms and simplified exceptions.
Added a descendant of Query, GvQuery, that tracks the google id.
This way the query can be eaisly reassociated with its source
if it is passed back to the GvConnection.  This will come in handy
for deleting queries from the server, which is a privacy thing.

Also, deleted unnecessary exceptions, and generally cleaned up the
SmsConnection interface.

TODO: Move all gv stuff into its own subpackage of sms.