PioText now ignores queries over 1/2 hour old.
Upon recieving a new query, PioText checks the send time against
the current system time, and ignores it if more than a half
hour has past.  This will prevent PioTxt from spamming users
upon startup.

For instance, if the server crashed for a day, a bunch of messages
would build up on GV.  When restarted the next day, people aren't
going to want their queries responded to from the previous day.
That would be annoying junk.  Even a half hour is pushing it.