Fixed bug in message generation.
If a message handler was passed a query from the distant past (or
the future, for that matter) the message handler would generate a
"(service break)" warning before listing any stops.

The test case caught it, and now stop lookups should be entirely
independant of the current time.  They should be based only
on the recieved time.