Trashed homebrew xml parser for lib suppored xml.
Google Voice Java now has a nice suite of utilities for pulling sms
out of the xml/json mix that google delivers.  Instead of the
previous homebrew solution for parsing, we now use the library.

However, this has added some new dependencies... which sucks.
They use a whole suite of utilites that must be included in lib/
now. We now have about 2 megs of .jars that we depend on...
which is silly.

Is it possible that jaxen, jtidy, and dom4j are already in the
compiled google-voice-137.jar or that we could do a custom build
with them in it?  That might cut down dependency bloat.