First ant build file.
Piotxt now has a build file.  However...

The junit tests are rampantly failing because ant doesn't know
how to properly ignore abstract test cases.  The fix for this
is probably to have abstract in their name, and add an ignore
to the fileset junit runs on.

As well, the build target still has library references for my
machine, and is just not well formed yet.  The whole jar-within-a-
jar thing is quite funky, and probably unnecessary.  Hopefully
we'll be able to just rip out all the .class files we need
and package them within piotxt.jar, instead of containing the
entire library of jars within it.

Also, compile is recompiling all the sources instead of just the
changed ones.  I know ant is smarter than that... I just need to
figure out how to keep the unchanged files.